fabrics.fabric_delete: 500 error on attempt to delete fabric containing switch(es)
1. app/endpoints/lan_fabric/rest/control/fabrics/fabric_delete.py

- Return a 500 response when user tries to delete a fabric that contains switch(es).
- The 500 response matches NDFC's

2. app/endpoints/lan_fabric/rest/control/fabrics/inventory/common.py

- build_db_switch()
  - Copy function from app/endpoints/lan_fabric/rest/control/fabrics/inventory/discover_post.py

3. tests/unit/v1/test_v1_fabric.py

- Add test case test_v1_fabric_delete_120
  - Attempt to delete a fabric that contains switches
  - Leverage build_db_switch() from 2 above

- In all other test cases, rename FABRIC_NAME from "f1" to "F1"

+++ app/v1/endpoints/lan_fabric/rest/control/fabrics/fabric_delete.py
@@ -3,6 +3,7 @@ from sqlmodel import Session, select
 
 from .......db import get_session
 from ......models.fabric import Fabric
+from ......models.inventory import SwitchDbModel
 from .common import build_404_response
 
 router = APIRouter(
@@ -25,6 +26,12 @@ def v1_fabric_delete(*, session: Session = Depends(get_session), fabric_name: st
     if not db_fabric:
         path = f"{router.prefix}/{fabric_name}"
         raise HTTPException(status_code=404, detail=build_404_response(path))
+    fabric_id = db_fabric.id
+    db_switches = session.exec(select(SwitchDbModel).where(SwitchDbModel.fabricId == fabric_id)).all()
+    if len(db_switches) != 0:
+        msg = "Failed to delete the fabric. Please check Events for possible reasons."
+        raise HTTPException(status_code=500, detail=msg)
+
     session.delete(db_fabric)
     session.commit()
     return {f"Fabric '{fabric_name}' is deleted successfully!"}
